Update to work with Aluminium GA
This updates pointers to Aluminum versions, adding porting changes.

We disable the odl-l2switch-ui feature, as we do not have dlux
available. It should be pretty easy to add back.

Places where we need to change logic due to Binding mapping having been
changed receive a bit more substantial changes in that:
- we use try-with-resources blocks for read transactions
- dispatch logic is updated to reduce unnecessary nesting
- a few places are also updated to use Uint types

// FIXME: audit code changes

+++ l2switch-main/src/test/java/org/opendaylight/l2switch/util/InstanceIdentifierUtilsTest.java
@@ -25,10 +25,11 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.node.No
 import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.nodes.Node;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.inventory.rev130819.nodes.NodeKey;
 import org.opendaylight.yangtools.yang.binding.InstanceIdentifier;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 public class InstanceIdentifierUtilsTest {
 
-    private static final Short NUM_ID_1 = 1;
+    private static final Uint8 NUM_ID_1 = Uint8.valueOf(1);
     private static final String STR_ID_1 = "id1";
 
     @Test
